+++ apps/web/lib/api/server/v1/userBike.ts
@@ -3,8 +3,6 @@ import { prisma } from '@repo/database'
 import {
   ApiResponseUserBikeList,
   ApiResponseUserBikeDetail,
-  ApiResponseFuelLogDetail,
-  ApiResponseFuelLogList,
   ApiResponseFuelInsight,
   ApiResponseMaintenanceLogDetail,
   ApiResponseMaintenanceLogList,
@@ -31,11 +29,6 @@ import {
   UserBikeUpdateRequestSchema,
   UserBikeListQuerySchema,
   FuelInsightQuerySchema,
-  FuelLogRegisterRequestSchema,
-  FuelLogUpdateRequestSchema,
-  FuelLogDeleteRequestSchema,
-  FuelLogListQuerySchema,
-  FuelLogDetailParamSchema,
   TouringRegisterRequestSchema,
   TouringStartEndRequestSchema,
   TouringListQuerySchema,
@@ -49,11 +42,7 @@ import {
 import { ApiV1Error } from '../errors/ApiV1Error'
 import { MyUserBikeDetail } from '../interfaces/IMyUserBikeRepository'
 import { honoAuthMiddleware } from '../middlewares/honoAuth'
-import {
-  zodValidateJson,
-  zodValidateParam,
-  zodValidateQuery,
-} from '../middlewares/zodValidation'
+import { zodValidateJson, zodValidateQuery } from '../middlewares/zodValidation'
 import { PrismaBikeRepository } from '../repositories/PrismaBikeRepository'
 import { PrismaFuelInsightRepository } from '../repositories/PrismaFuelInsightRepository'
 import { PrismaFuelLogRepository } from '../repositories/PrismaFuelLogRepository'
@@ -64,7 +53,6 @@ import { PrismaSpotRepository } from '../repositories/PrismaSpotRepository'
 import { PrismaTouringRepository } from '../repositories/PrismaTouringRepository'
 import { PrismaUserBikeRepository } from '../repositories/PrismaUserBikeRepository'
 import { FuelInsightService } from '../services/FuelInsightService'
-import { FuelLogService } from '../services/FuelLogService'
 import { MaintenanceLogService } from '../services/MaintenanceLogService'
 import { SpotService } from '../services/SpotService'
 import { TouringService } from '../services/TouringService'
@@ -73,6 +61,7 @@ import { FuelInsightSearchParams } from '../valueObjects/FuelInsightSearchParams
 import { FuelLogSearchParams } from '../valueObjects/FuelLogSearchParams'
 import { TouringSearchParams } from '../valueObjects/TouringSearchParams'
 import { UserBikeSearchParams } from '../valueObjects/UserBikeSearchParams'
+import userBikeFuelLogs from './userBike/fuelLogs'
 
 const userBike = new Hono()
 
@@ -537,287 +526,7 @@ userBike.get('/bike/:myUserBikeId/history', honoAuthMiddleware, async (c) => {
   })
 })
 
-userBike.get(
-  '/bike/:myUserBikeId/fuel-logs',
-  honoAuthMiddleware,
-  zodValidateQuery(FuelLogListQuerySchema),
-  async (c) => {
-    const { userId } = c.var.user!
-    const myUserBikeId = c.req.param('myUserBikeId')
-    const query = c.req.valid('query')
-
-    const searchParams = new FuelLogSearchParams({
-      page: query.page,
-      pageSize: query['per-size'],
-      sortBy: query['sort-by'] === 'mileage' ? 'mileage' : 'refueledAt',
-      sortOrder: query['sort-order'],
-      period: query.period,
-      startDate: query.startDate,
-      endDate: query.endDate,
-    })
-
-    const fuelLogRepo = new PrismaFuelLogRepository(prisma)
-    const myUserBikeRepo = new PrismaMyUserBikeRepository(prisma)
-    const userBikeRepo = new PrismaUserBikeRepository(prisma)
-    const touringRepo = new PrismaTouringRepository(prisma)
-    const fuelLogService = new FuelLogService(
-      fuelLogRepo,
-      myUserBikeRepo,
-      userBikeRepo,
-      touringRepo
-    )
-
-    const fuelLogs = await fuelLogService.getFuelLogs(
-      createMyUserBikeId(myUserBikeId),
-      createUserId(userId),
-      searchParams
-    )
-
-    return c.json<SuccessResponse<ApiResponseFuelLogList>>(
-      {
-        status: 'success',
-        data: fuelLogs.map((log) => {
-          return {
-            fuelLogId: log.id,
-            refueledAt: log.refueledAt.toISOString(),
-            mileage: log.mileage,
-            previousMileage: log.previousMileage,
-            amount: log.amount,
-            totalPrice: log.totalPrice,
-            memo: log.memo,
-            fuelEfficiency: log.fuelEfficiency,
-            pricePerLiter: log.pricePerLiter,
-            touringId: log.touringId,
-            touringTitle: log.touringTitle,
-          }
-        }),
-        message: '燃料ログ一覧取得成功',
-      },
-      200
-    )
-  }
-)
-
-userBike.get(
-  '/bike/:myUserBikeId/fuel-logs/:fuelLogId',
-  honoAuthMiddleware,
-  zodValidateParam(FuelLogDetailParamSchema),
-  async (c) => {
-    const { userId } = c.var.user!
-    const params = c.req.valid('param')
-
-    const fuelLogRepo = new PrismaFuelLogRepository(prisma)
-    const myUserBikeRepo = new PrismaMyUserBikeRepository(prisma)
-    const userBikeRepo = new PrismaUserBikeRepository(prisma)
-    const touringRepo = new PrismaTouringRepository(prisma)
-    const fuelLogService = new FuelLogService(
-      fuelLogRepo,
-      myUserBikeRepo,
-      userBikeRepo,
-      touringRepo
-    )
-
-    const fuelLog = await fuelLogService.getFuelLogDetail(
-      createFuelLogId(params.fuelLogId),
-      createMyUserBikeId(params.myUserBikeId),
-      createUserId(userId)
-    )
-
-    return c.json<SuccessResponse<ApiResponseFuelLogDetail>>({
-      status: 'success',
-      data: {
-        fuelLogId: fuelLog.id,
-        refueledAt: fuelLog.refueledAt.toISOString(),
-        mileage: fuelLog.mileage,
-        previousMileage: fuelLog.previousMileage,
-        amount: fuelLog.amount,
-        totalPrice: fuelLog.totalPrice,
-        memo: fuelLog.memo,
-        fuelEfficiency: fuelLog.fuelEfficiency,
-        pricePerLiter: fuelLog.pricePerLiter,
-        touringId: fuelLog.touringId,
-        touringTitle: fuelLog.touringTitle,
-      },
-      message: '燃料ログ詳細取得成功',
-    })
-  }
-)
-
-userBike.post(
-  '/bike/:myUserBikeId/fuel-logs',
-  honoAuthMiddleware,
-  zodValidateJson(FuelLogRegisterRequestSchema),
-  async (c) => {
-    const { userId, role } = c.var.user!
-    const myUserBikeId = c.req.param('myUserBikeId')
-    const body = c.req.valid('json')
-
-    const result = await prisma.$transaction(async (t) => {
-      const fuelLogRepo = new PrismaFuelLogRepository(t)
-      const myUserBikeRepo = new PrismaMyUserBikeRepository(t)
-      const userBikeRepo = new PrismaUserBikeRepository(t)
-      const touringRepo = new PrismaTouringRepository(t)
-      const service = new FuelLogService(
-        fuelLogRepo,
-        myUserBikeRepo,
-        userBikeRepo,
-        touringRepo
-      )
-
-      const fuelLog = await service.registerFuelLog({
-        myUserBikeId: createMyUserBikeId(myUserBikeId),
-        userId: createUserId(userId),
-        role,
-        refueledAt: body.refueledAt,
-        mileage: body.mileage,
-        previousMileage: body.previousMileage,
-        amount: body.amount,
-        totalPrice: body.totalPrice,
-        memo: body.memo,
-        updateTotalMileage: body.updateTotalMileage,
-      })
-
-      const historyRepo = new PrismaHistoryRepository(t)
-      await historyRepo.createHistory({
-        userId: createUserId(userId),
-        userMyBikeId: createMyUserBikeId(myUserBikeId),
-        type: 'FUEL_LOG',
-        occurredAt: fuelLog.refueledAt,
-        fuelLogId: fuelLog.id,
-      })
-
-      return fuelLog
-    })
-
-    return c.json<SuccessResponse<ApiResponseFuelLogDetail>>(
-      {
-        status: 'success',
-        data: {
-          fuelLogId: result.id,
-          refueledAt: result.refueledAt.toISOString(),
-          mileage: result.mileage,
-          previousMileage: result.previousMileage,
-          amount: result.amount,
-          totalPrice: result.totalPrice,
-          memo: result.memo,
-          fuelEfficiency: result.fuelEfficiency,
-          pricePerLiter: result.pricePerLiter,
-          touringId: result.touringId,
-          touringTitle: result.touringTitle,
-        },
-        message: '燃料ログ登録成功',
-      },
-      201
-    )
-  }
-)
-
-userBike.patch(
-  '/bike/:myUserBikeId/fuel-logs',
-  honoAuthMiddleware,
-  zodValidateJson(FuelLogUpdateRequestSchema),
-  async (c) => {
-    const { userId } = c.var.user!
-    const myUserBikeId = c.req.param('myUserBikeId')
-    const body = c.req.valid('json')
-
-    const result = await prisma.$transaction(async (t) => {
-      const fuelLogRepo = new PrismaFuelLogRepository(t)
-      const myUserBikeRepo = new PrismaMyUserBikeRepository(t)
-      const userBikeRepo = new PrismaUserBikeRepository(t)
-      const touringRepo = new PrismaTouringRepository(t)
-      const historyRepo = new PrismaHistoryRepository(t)
-      const service = new FuelLogService(
-        fuelLogRepo,
-        myUserBikeRepo,
-        userBikeRepo,
-        touringRepo
-      )
-
-      const updated = await service.updateFuelLog({
-        fuelLogId: createFuelLogId(body.fuelLogId),
-        myUserBikeId: createMyUserBikeId(myUserBikeId),
-        userId: createUserId(userId),
-        refueledAt: body.refueledAt,
-        mileage: body.mileage,
-        previousMileage: body.previousMileage,
-        amount: body.amount,
-        totalPrice: body.totalPrice,
-        memo: body.memo,
-      })
-
-      // 給油日時が変更された場合、ヒストリーの occurredAt を更新
-      if (body.refueledAt !== undefined) {
-        await historyRepo.updateOccurredAtByFuelLogId(
-          createFuelLogId(body.fuelLogId),
-          updated.refueledAt
-        )
-      }
-
-      return updated
-    })
-
-    return c.json<SuccessResponse<ApiResponseFuelLogDetail>>(
-      {
-        status: 'success',
-        data: {
-          fuelLogId: result.id,
-          refueledAt: result.refueledAt.toISOString(),
-          mileage: result.mileage,
-          previousMileage: result.previousMileage,
-          amount: result.amount,
-          totalPrice: result.totalPrice,
-          memo: result.memo,
-          fuelEfficiency: result.fuelEfficiency,
-          pricePerLiter: result.pricePerLiter,
-          touringId: result.touringId,
-          touringTitle: result.touringTitle,
-        },
-        message: '燃料ログ更新成功',
-      },
-      200
-    )
-  }
-)
-
-userBike.delete(
-  '/bike/:myUserBikeId/fuel-logs',
-  honoAuthMiddleware,
-  zodValidateJson(FuelLogDeleteRequestSchema),
-  async (c) => {
-    const { userId } = c.var.user!
-    const myUserBikeId = c.req.param('myUserBikeId')
-    const body = c.req.valid('json')
-
-    await prisma.$transaction((t) => {
-      const fuelLogRepo = new PrismaFuelLogRepository(t)
-      const myUserBikeRepo = new PrismaMyUserBikeRepository(t)
-      const userBikeRepo = new PrismaUserBikeRepository(t)
-      const touringRepo = new PrismaTouringRepository(t)
-      const service = new FuelLogService(
-        fuelLogRepo,
-        myUserBikeRepo,
-        userBikeRepo,
-        touringRepo
-      )
-
-      return service.deleteFuelLog({
-        fuelLogId: createFuelLogId(body.fuelLogId),
-        myUserBikeId: createMyUserBikeId(myUserBikeId),
-        userId: createUserId(userId),
-      })
-    })
-
-    return c.json<SuccessResponse<undefined>>(
-      {
-        status: 'success',
-        message: '燃料ログ削除成功',
-        data: undefined,
-      },
-      200
-    )
-  }
-)
+userBike.route('/', userBikeFuelLogs)
 
 userBike.delete(
   '/bike/:myUserBikeId/tourings',
